refactor: profile 관련 sonarqube 문제점 해결
